+++ app/src/main/java/com/diary/paintlog/utils/listener/BaseDateListener.kt
@@ -1,0 +1,8 @@
+package com.diary.paintlog.utils.listener
+
+import java.time.LocalDateTime
+
+interface BaseDateListener {
+
+    fun onItemClick(baseDate: LocalDateTime)
+}
