+++ app/src/main/java/com/diary/paintlog/utils/listener/DataListener.kt
@@ -1,4 +1,4 @@
-package com.diary.paintlog.utils
+package com.diary.paintlog.utils.listener
 
 interface DataListener {
     fun onDataReceived(data: Map<String, String>)
