+++ app/src/main/java/com/diary/paintlog/utils/listener/DiaryIdListener.kt
@@ -1,4 +1,4 @@
-package com.diary.paintlog.utils
+package com.diary.paintlog.utils.listener
 
 interface DiaryIdListener {
     fun onItemClick(diaryId: Long)
